requirements.txt
# Web アプリケーション
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0

# ブラウザ自動操作
playwright==1.41.2

# AI 画像処理・テキスト生成
google-generativeai==0.3.2
Pillow==10.1.0

# 画像処理
pybase64==1.3.1

# テスト
pytest==7.4.3
pytest-flask==1.3.0
coverage==7.3.2
